Create Project 1
This is an online retail dataset and In this project, I aimed to clean a dataset by removing unnecessary information and ensuring data quality. Specifically, I removed columns where the customer ID is null, the quantity is negative, and the price is less than 0. Then, I removed all duplicates and created a temporary table to store the cleaned data. The code provided in this file demonstrates how I achieved these steps. Feel free to review and provide feedback